# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * EduWhiteboard API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LessonStatus = typeof LessonStatus[keyof typeof LessonStatus];


export const LessonStatus = {
  in_progress: 'in_progress',
  completed: 'completed',
  pending: 'pending',
  rendering: 'rendering',
} as const;

export interface Lesson {
  id: number;
  title: string;
  grade: string;
  language: string;
  /** @nullable */
  durationMinutes?: number | null;
  status: LessonStatus;
  /** @nullable */
  thumbnailUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type LessonInputStatus = typeof LessonInputStatus[keyof typeof LessonInputStatus];


export const LessonInputStatus = {
  in_progress: 'in_progress',
  completed: 'completed',
  pending: 'pending',
  rendering: 'rendering',
} as const;

export interface LessonInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  grade: string;
  /** @minLength 1 */
  language: string;
  /** @nullable */
  durationMinutes?: number | null;
  status?: LessonInputStatus;
  /** @nullable */
  thumbnailUrl?: string | null;
}

export type LessonUpdateStatus = typeof LessonUpdateStatus[keyof typeof LessonUpdateStatus];


export const LessonUpdateStatus = {
  in_progress: 'in_progress',
  completed: 'completed',
  pending: 'pending',
  rendering: 'rendering',
} as const;

export interface LessonUpdate {
  /** @minLength 1 */
  title?: string;
  grade?: string;
  language?: string;
  /** @nullable */
  durationMinutes?: number | null;
  status?: LessonUpdateStatus;
  /** @nullable */
  thumbnailUrl?: string | null;
}

export type LessonStatsByStatus = {
  in_progress: number;
  completed: number;
  pending: number;
  rendering: number;
};

export interface LessonStats {
  total: number;
  byStatus: LessonStatsByStatus;
}

export interface ErrorResponse {
  error: string;
}

export type GetLessonsParams = {
/**
 * Search lessons by title
 */
search?: string;
/**
 * Filter by status
 */
status?: GetLessonsStatus;
};

export type GetLessonsStatus = typeof GetLessonsStatus[keyof typeof GetLessonsStatus];


export const GetLessonsStatus = {
  in_progress: 'in_progress',
  completed: 'completed',
  pending: 'pending',
  rendering: 'rendering',
} as const;