# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * EduWhiteboard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LessonStatus } from './lessonStatus';

export interface Lesson {
  id: number;
  title: string;
  grade: string;
  language: string;
  /** @nullable */
  durationMinutes?: number | null;
  status: LessonStatus;
  /** @nullable */
  thumbnailUrl?: string | null;
  createdAt: Date;
  updatedAt: Date;
}