# lib/api-zod/src/generated/types/lessonInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * EduWhiteboard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LessonInputStatus } from './lessonInputStatus';

export interface LessonInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  grade: string;
  /** @minLength 1 */
  language: string;
  /** @nullable */
  durationMinutes?: number | null;
  status?: LessonInputStatus;
  /** @nullable */
  thumbnailUrl?: string | null;
}