# lib/api-zod/src/generated/types/lessonStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * EduWhiteboard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LessonStatsByStatus } from './lessonStatsByStatus';

export interface LessonStats {
  total: number;
  byStatus: LessonStatsByStatus;
}