# lib/api-zod/src/generated/types/lessonUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * EduWhiteboard API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LessonUpdateStatus } from './lessonUpdateStatus';

export interface LessonUpdate {
  /** @minLength 1 */
  title?: string;
  grade?: string;
  language?: string;
  /** @nullable */
  durationMinutes?: number | null;
  status?: LessonUpdateStatus;
  /** @nullable */
  thumbnailUrl?: string | null;
}